+++ app/src/main/java/de/thm/lampgame/controller/item/ItemController.kt
@@ -4,5 +4,5 @@ import de.thm.lampgame.model.item.ItemModel
 
 abstract class ItemController {
     open lateinit var itemModel: ItemModel
-    abstract fun draw(canvas: Any, velocity: Int)
+    abstract fun draw(canvas: Any, velocity: Double)
 }
