+++ app/src/main/java/de/thm/lampgame/controller/obstacles/BitmapBouncingSaw.kt
@@ -8,7 +8,7 @@ import de.thm.lampgame.R
 import de.thm.lampgame.model.obstacles.ObstacleModel
 import de.thm.lampgame.model.obstacles.ObstacleNames
 
-class BitmapBouncingSaw(context: Context, width: Int, height: Int, x: Int, y: Int) : ObstacleController() {
+class BitmapBouncingSaw(context: Context, width: Int, height: Int, x: Double, y: Double) : ObstacleController() {
     // creates the associated model
     override var obstacleModel = ObstacleModel(
         ObstacleNames.BOUNCINGSAW,
@@ -38,7 +38,7 @@ class BitmapBouncingSaw(context: Context, width: Int, height: Int, x: Int, y: In
 
     // change the x coordinate by the given velocity value and draw the obstacle
     // -> get the required data from the associated model
-    override fun draw(canvas: Any, velocityX: Int, velocityY: Int) {
+    override fun draw(canvas: Any, velocityX: Double, velocityY: Double) {
         obstacleModel.changeableX -= velocityX
         obstacleModel.changeYCoords(velocityY)
