+++ app/src/main/java/de/thm/lampgame/controller/obstacles/BitmapTerrain.kt
@@ -9,7 +9,7 @@ import de.thm.lampgame.model.obstacles.ObstacleModel
 import de.thm.lampgame.model.obstacles.ObstacleNames
 
 
-class BitmapTerrain(context: Context, width: Int, height: Int, x: Int, y: Int) : ObstacleController() {
+class BitmapTerrain(context: Context, width: Int, height: Int, x: Double, y: Double) : ObstacleController() {
     // creates the associated model
     override var obstacleModel = ObstacleModel(ObstacleNames.TERRAIN, width, height, x, y, false)
 
@@ -31,7 +31,7 @@ class BitmapTerrain(context: Context, width: Int, height: Int, x: Int, y: Int) :
 
     // change the x coordinate by the given velocity value and draw the obstacle
     // -> get the required data from the associated model
-    override fun draw(canvas: Any, velocityX: Int, velocityY: Int) {
+    override fun draw(canvas: Any, velocityX: Double, velocityY: Double) {
         obstacleModel.changeableX -= velocityX
         (canvas as Canvas).drawBitmap(
             obstacleModel.bmp as Bitmap,
