+++ app/src/main/java/de/thm/lampgame/controller/obstacles/BitmapWater.kt
@@ -8,14 +8,14 @@ import de.thm.lampgame.R
 import de.thm.lampgame.model.obstacles.ObstacleModel
 import de.thm.lampgame.model.obstacles.ObstacleNames
 
-class BitmapWater(context: Context, width: Int, height: Int, x: Int, y: Int) : ObstacleController() {
+class BitmapWater(context: Context, width: Int, height: Int, x: Double, y: Double) : ObstacleController() {
     // creates the associated model
     override var obstacleModel = ObstacleModel(
         ObstacleNames.WATER,
         width,
         (0.13 * height).toInt(),
         x,
-        (0.87 * y).toInt(),
+        (0.87 * y),
         true
     )
 
@@ -37,7 +37,7 @@ class BitmapWater(context: Context, width: Int, height: Int, x: Int, y: Int) : O
 
     // change the x coordinate by the given velocity value and draw the obstacle
     // -> get the required data from the associated model
-    override fun draw(canvas: Any, velocityX: Int, velocityY: Int) {
+    override fun draw(canvas: Any, velocityX: Double, velocityY: Double) {
         obstacleModel.changeableX -= velocityX
         (canvas as Canvas).drawBitmap(
             obstacleModel.bmp as Bitmap,
