+++ app/src/main/java/de/thm/lampgame/controller/obstacles/ObstacleController.kt
@@ -4,5 +4,5 @@ import de.thm.lampgame.model.obstacles.ObstacleModel
 
 abstract class ObstacleController {
     open lateinit var obstacleModel: ObstacleModel
-    abstract fun draw(canvas: Any, velocityX: Int, velocityY: Int)
+    abstract fun draw(canvas: Any, velocityX: Double, velocityY: Double)
 }
