+++ app/src/main/java/de/thm/lampgame/model/item/TorchModel.kt
@@ -2,7 +2,7 @@ package de.thm.lampgame.model.item
 
 import de.thm.lampgame.controller.Player
 
-class TorchModel(screenHeight: Int, screenWidth: Int, height: Int, width: Int, x: Int, y: Int) :
+class TorchModel(screenHeight: Int, screenWidth: Int, height: Int, width: Int, x: Double, y: Double) :
     ItemModel(screenHeight, screenWidth, height, width, x, y) {
 
     // determines the effect of the item
